add SyncMaster unitest
